Add group directory in src/components/header
