+++ src/components/header/Header.tsx
@@ -14,9 +14,8 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import EventAvailableIcon from '@material-ui/icons/EventAvailable';
 import { logOut } from '../../reducks/users/operations';
-import { MobileDrawer } from './index';
-import { InvitationNotifications } from '../todo';
-import SwitchEntity from './SwitchEntity';
+import { MobileDrawer, InvitationNotifications } from './index';
+import { SwitchEntity } from './group';
 import { getPathGroupId, getPathTemplateName } from '../../lib/path';
 import { fetchUserInfo } from '../../reducks/users/operations';
 import { fetchGroups } from '../../reducks/groups/operations';
