+++ src/components/header/group/index.ts
@@ -1,0 +1,5 @@
+export { default as CreateGroups } from './CreateGroups';
+export { default as EditGroupMembers } from './EditGroupMembers';
+export { default as EditGroupName } from './EditGroupName';
+export { default as GroupMenuButton } from './GroupMenuButton';
+export { default as SwitchEntity } from './SwitchEntity';
