+++ src/components/header/index.ts
@@ -1,3 +1,3 @@
 export { default as Header } from './Header';
 export { default as MobileDrawer } from './MobileDrawer';
-export { default as SwitchEntity } from './SwitchEntity';
+export { default as InvitationNotifications } from './InvitationNotifications';
