+++ src/components/todo/SearchField.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import { DatePicker, SelectCompleteFlag, SelectLimit, SelectSortType } from '../uikit';
-import { SelectDateType, SelectUsers, TextInput } from './index';
+import { SelectDateType, TextInput } from './index';
 import '../../assets/todo/search-field.scss';
 import { useDispatch } from 'react-redux';
 import { searchTodoList } from '../../reducks/todoList/operations';
 import { searchTodoRequestData } from '../../reducks/todoList/types';
-import { Groups } from '../../reducks/groups/types';
 import { searchGroupTodoRequestData } from '../../reducks/groupTodoList/types';
 import { useLocation, useParams } from 'react-router';
 import { searchGroupTodoList } from '../../reducks/groupTodoList/operations';
 
 interface SearchFieldProps {
-  approvedGroups: Groups;
   closeSearch: () => void;
   setCurrentDateType: React.Dispatch<React.SetStateAction<string>>;
   setOpenSearchResultTodoList: React.Dispatch<React.SetStateAction<boolean>>;
@@ -154,14 +152,6 @@ const SearchField = (props: SearchFieldProps) => {
           </div>
         );
       })}
-      {pathName === 'group' && (
-        <div className="search-field__select-contents">
-          <span className="search-field__select-contents--key">ユーザーを選択</span>
-          <span className="search-field__select-contents--value">
-            <SelectUsers approvedGroups={props.approvedGroups} setUsersIds={props.setUsersIds} />
-          </span>
-        </div>
-      )}
       <div className="search-field__search-btn">
         <button
           className="save-btn"
