+++ src/components/todo/index.tsx
@@ -1,13 +1,7 @@
 export { default as AddTodo } from './AddTodo';
-export { default as CreateGroups } from './CreateGroups';
-export { default as EditGroupMembers } from './EditGroupMembers';
-export { default as EditGroupName } from './EditGroupName';
 export { default as ExistsTodoLists } from './ExistsTodoLists';
 export { default as ExpiredTodoList } from './ExpiredTodoList';
-export { default as GroupMenuButton } from './GroupMenuButton';
-export { default as GroupName } from './GroupName';
 export { default as InputTodoList } from './InputTodoList';
-export { default as InvitationNotifications } from './InvitationNotifications';
 export { default as MobileTodoMenu } from './MobileTodoMenu';
 export { default as MonthlyTodoList } from './MonthlyTodoList';
 export { default as SearchField } from './SearchField';
